reuse build_magnet
